FileReader Error Handling Unit Test:   - Test ServerStartError   - Move FileIO mocks into separate files.
