+++ Tests/FileIOTests/FileWriterTest.swift
@@ -1,16 +1,6 @@
 import XCTest
 @testable import FileIO
 
-class MockContent: Writable {
-  var didWrite: Bool = false
-  var givenArgs: URL = URL(fileURLWithPath: "")
-
-  func write<WriteableLocation>(to destination: WriteableLocation) throws {
-    didWrite = true
-    givenArgs = destination as! URL
-  }
-}
-
 class FileWriterTest: XCTestCase {
   func testInitDoesNotCallWritableWriteMethod() {
     let path = "/some/path/to/somewhere"
