+++ Tests/FileIOTests/MockFileManager.swift
@@ -1,0 +1,7 @@
+import FileIO
+
+class MockFileManager: FileIO {
+  func contentsOfDirectory(atPath path: String) throws -> [String] {
+    return ["file1", "file2"]
+  }
+}
